% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize.R
\name{discretize}
\alias{discretize}
\title{Discretize two numeric columns of a data frame into a matrix.}
\usage{
discretize(.data, dims = c(600, 600), x = "x", y = "y")
}
\arguments{
\item{.data}{Data frame}

\item{dims}{Dimensions of matrix}

\item{x}{Name of x column in the data frame}

\item{y}{Name of y column in the data frame}
}
\value{
A matrix with dimensions \code{dims}
}
\description{
Discretize two numeric columns of a data frame into a matrix.
}
