% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recolour.R
\name{recolour}
\alias{recolour}
\title{Use palette from ColorBrewer to recolour the image}
\usage{
recolour(x, palette = "Blues", zero_colour = "#FFFFFF",
  invert = FALSE)
}
\arguments{
\item{x}{Matrix}

\item{palette}{ColorBrewer palette}

\item{zero_colour}{Colour to use for zero and NA values}
}
\value{
Matrix
}
\description{
Use palette from ColorBrewer to recolour the image
}
