% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strange_attractor.R
\name{strange_attractor}
\alias{strange_attractor}
\title{Generate strange attractor.}
\usage{
strange_attractor(a, n, x0 = 1, y0 = 1)
}
\arguments{
\item{a}{Numeric vector of length 14}

\item{n}{Number of points to generate}

\item{x0}{Initial value for x}

\item{y0}{Initial value for y}
}
\value{
Data frame with columns \code{x} and \code{y}, and \code{n} rows.
}
\description{
Generate strange attractor.
}
