% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_quantiles.R
\name{trim_quantiles}
\alias{trim_quantiles}
\title{Trim x and y range based on quantiles}
\usage{
trim_quantiles(x, q = 0.05)
}
\arguments{
\item{x}{A data frame with columns x and y}

\item{q}{Quantiles cutoff}
}
\value{
A data frame
}
\description{
Trim x and y range based on quantiles
}
