% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recolour.R
\name{recolour}
\alias{recolour}
\alias{recolor}
\title{Use palette from ColorBrewer to recolour the image}
\usage{
recolour(x, palette = "Blues", zero_colour = "#FFFFFF",
  invert = FALSE)

recolor(x, palette = "Blues", zero_colour = "#FFFFFF",
  invert = FALSE)
}
\arguments{
\item{x}{Matrix}

\item{palette}{ColorBrewer palette. This value is passed to \code{\link[scales:col_numeric]{scales::col_numeric()}}}

\item{zero_colour}{Colour to use for zero and NA values}

\item{invert}{If \code{TRUE}, inverts the colour scale}
}
\value{
Matrix
}
\description{
Use palette from ColorBrewer to recolour the image
}
