% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize.R
\name{strange_attractor}
\alias{strange_attractor}
\title{Generate strange attractor.}
\usage{
strange_attractor(a, n, x0 = 0, y0 = 0)
}
\arguments{
\item{a}{Numeric vector of length 14}

\item{n}{Number of points to generate}

\item{x0}{Initial value for x}

\item{y0}{Initial value for y}
}
\description{
Generate strange attractor.
}
